// frontend/src/backend.d.ts
export enum BookingStatus {
    cancelled = "cancelled",
    pending = "pending",
    confirmed = "confirmed"
}
export enum PaymentPlan {
    cod = "cod",
    partialUpfront = "partialUpfront",
    depositRequired = "depositRequired",
    fullUpfront = "fullUpfront"
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export enum Variant_new_closed_reviewed {
    new_ = "new",
    closed = "closed",
    reviewed = "reviewed"
}
export enum Variant_new_verified_rejected {
    new_ = "new",
    verified = "verified",
    rejected = "rejected"
}